codeql issue fixed

+++ nitrite-jackson-mapper/src/test/java/org/dizitart/no2/test/transaction/TransactionRepositoryTest.java
@@ -558,7 +558,7 @@ public class TransactionRepositoryTest extends BaseObjectRepositoryTest {
         try (Session session = db.createSession()) {
             ExecutorService executorService = Executors.newCachedThreadPool();
             for (int i = 0; i < 10; i++) {
-                final int fi = i;
+                final long fi = i;
                 Future<?> future = executorService.submit(() -> {
                     Transaction transaction = session.beginTransaction();
                     try {
@@ -604,7 +604,7 @@ public class TransactionRepositoryTest extends BaseObjectRepositoryTest {
         try (Session session = db.createSession()) {
             ExecutorService executorService = Executors.newCachedThreadPool();
             for (int i = 0; i < 10; i++) {
-                final int fi = i;
+                final long fi = i;
                 Future<?> future = executorService.submit(() -> {
                     Transaction transaction = session.beginTransaction();
                     try {
